% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOMED_Helper.R
\name{calculate_cell_proportions}
\alias{calculate_cell_proportions}
\title{Calculate Cell Type Proportions Using Deconvolution Method}
\usage{
calculate_cell_proportions(beta, reference, celltype)
}
\arguments{
\item{beta}{A matrix of methylation data (CpG by sample).}

\item{reference}{A reference matrix (cell type-specific methylation profiles).}

\item{celltype}{A character string specifying the cell type for which the proportion is calculated.}
}
\value{
A numeric vector of predicted cell proportions for the specified cell type.
}
\description{
This function calculates the cell type proportion for a given 2-level structured cell type using a deconvolution method.
It uses the `epidish` package to perform deconvolution and returns the estimated cell proportion for a specified cell type.
}
\examples{
\dontrun{
cell_proportion <- calculate_cell_proportions(beta_data, reference_data)
}
}
