% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOMED_Helper.R
\name{deconv_limma}
\alias{deconv_limma}
\title{Perform Limma-based Probe Selection for Deconvolution}
\usage{
deconv_limma(beta, pd, celltypes, p_val = 1e-11, n_prob = 500, by = "logfc")
}
\arguments{
\item{beta}{Methylation beta matrix (CpG by sample).}

\item{pd}{A data frame containing the cell type annotation, must have a column with name 'CellType'}

\item{celltypes}{A vector of cell types to be used for probe selection.}

\item{p_val}{The p-value threshold for significance in the differential expression analysis.}

\item{n_prob}{The number of probes to select.}

\item{by}{Specifies the method for probe selection, either by log fold change (`'logfc'`) or adjusted p-value (`'adj.P.Val'`).}
}
\value{
A vector of selected probe names.
}
\description{
This function uses the `limma` package to select probes based on differential expression between specified cell types.
}
\examples{
\dontrun{
selected_probes <- deconv_limma(beta,pd,celltypes,p_val= 1e-11,n_prob= 500,by='logfc')
}
}
