% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOMED_Estimate.R
\name{HOMED_Estimate}
\alias{HOMED_Estimate}
\title{Estimate Cell Type Proportions Using Hierarchical Deconvolution}
\usage{
HOMED_Estimate(celltypes, deconv_beta, reference_obj)
}
\arguments{
\item{celltypes}{A named list where each element contains the names of cell types at level 1 and their corresponding subtypes at level 2.
For example, `list("Tcell" = c("Treg", "Th1", "Th2"), "Bcell" = c("Naive", "Plasma"))`.}

\item{deconv_beta}{A matrix of methylation beta values (CpG by sample), which will be used to calculate the cell proportions.}

\item{reference_obj}{A list containing reference matrices for both level 1 and level 2 deconvolution.
`reference_obj$layer1_reference_list` and `reference_obj$layer2_reference_list` should each be a list where
each element corresponds to a reference matrix for a specific cell type.}
}
\value{
A list containing the estimated cell type proportions. The `deconvolution_proportion` element of the list contains
        the merged proportions for both level 1 and level 2 cell types.
}
\description{
This function performs hierarchical deconvolution by estimating cell type proportions at two levels (level 1 and level 2).
It first estimates the proportions for level 1 cell types, and then for level 2 cell types. The results from both levels are merged
to generate the final estimated proportions for each sample.
}
\examples{
\dontrun{
estimated_proportions <- HOMED_Estimate(celltypes, deconv_beta, reference_obj)
}
}
