% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOMED_Reference.R
\name{HOMED_Reference}
\alias{HOMED_Reference}
\title{Construct Hierarchical Reference Matrices for Cell Type Deconvolution}
\usage{
HOMED_Reference(
  FACS_beta,
  FACS_pd,
  training_beta,
  training_prop,
  celltypes,
  p_val = 0.05,
  numProbes = 500,
  maxIt = c(20, 20),
  libSize = 500,
  rmse_improve_thresh = 0.001
)
}
\arguments{
\item{FACS_beta}{A matrix of methylation data for the FACS samples (CpG by sample).}

\item{FACS_pd}{A data frame containing phenotype data for the FACS samples, including cell type annotations.}

\item{training_beta}{A matrix of training methylation data (CpG by sample) used for model optimization.}

\item{training_prop}{A matrix of cell type proportions (cell type by sample) for the training samples.}

\item{celltypes}{A named list of cell types, where each cell type is associated with a list of its subtypes at level 2.}

\item{p_val}{The p-value threshold for probe selection using `limma`.}

\item{numProbes}{The number of probes to select for each cell type.}

\item{maxIt}{A vector of two integers specifying the maximum number of iterations for level 1 and level 2 optimization.}

\item{libSize}{The size of the library to use for optimization.}

\item{rmse_improve_thresh}{The threshold for improvement in RMSE between iterations to stop the optimization.}
}
\value{
A list containing two elements:
\item{layer1_reference_list}{A list of reference matrices for level 1 cell types.}
\item{layer2_reference_list}{A list of reference matrices for level 2 cell types.}
}
\description{
This function creates hierarchical reference matrices for deconvolution by performing probe selection and optimization
of coefficients for both level 1 and level 2 cell types. It uses a combination of `limma` for probe selection and an
optimization method (IDOL) to refine the reference matrices based on training data.
}
\examples{
\dontrun{
references <- HOMED_Reference(FACS_beta,FACS_pd,
training_beta,training_prop,celltypes,p_val = 1e-11,numProbes = 500)
}
}
