% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOMED_Helper.R
\name{IDOL_opt_rmse}
\alias{IDOL_opt_rmse}
\title{Optimize Coefficients Using RMSE for Deconvolution}
\usage{
IDOL_opt_rmse(
  candDMRFinderObject,
  trainingBetas,
  trainingCovariates,
  libSize = 500,
  maxIt = 200,
  numCores = 4,
  IDOLobj = TRUE,
  rmse_improve_thresh = 1e-04,
  rmse_for = c("Hofbauer")
)
}
\arguments{
\item{candDMRFinderObject}{A list or object containing candidate probes and coefficient estimates.}

\item{trainingBetas}{A matrix of methylation values for training samples.}

\item{trainingCovariates}{A matrix or data frame of the training covariates (e.g., cell proportions).}

\item{libSize}{The size of the library to use for optimization.}

\item{maxIt}{The maximum number of iterations for the optimization.}

\item{numCores}{The number of CPU cores to use for parallel processing.}

\item{IDOLobj}{Logical; whether to use the IDOL optimization method.}

\item{rmse_improve_thresh}{The threshold for RMSE improvement to stop the optimization.}

\item{rmse_for}{The cell type to optimize RMSE for.}
}
\value{
A list containing the optimized probes and coefficients, as well as RMSE and R2 values for each iteration.
}
\description{
This function optimizes methylation deconvolution coefficients by minimizing the RMSE between predicted and observed proportions.
}
\examples{
\dontrun{
idol_results <- IDOL_opt_rmse(candDMRFinderObject,
trainingBetas, trainingCovariates, libSize = 500, maxIt = 200)
}
}
