% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOMED_Helper.R
\name{summarize_level1}
\alias{summarize_level1}
\title{Summarize Level 1 Cell Type Proportions}
\usage{
summarize_level1(cell_types_list, prop_data)
}
\arguments{
\item{cell_types_list}{A list where each element contains the names of subtypes corresponding to a level 1 cell type.}

\item{prop_data}{A data frame or matrix of cell proportions (samples by cell type).}
}
\value{
A data frame with summarized cell proportions for each level 1 cell type.
}
\description{
This function summarizes cell type proportions at level 1 by summing the proportions of subtypes for each cell type.
The resulting data frame contains the summed proportions for each cell type across all samples.
}
\examples{
\dontrun{
summarized_proportions <- summarize_level1(celltypes, prop_data)
}
}
