% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOMED_Helper.R
\name{HOMED_plot}
\alias{HOMED_plot}
\title{Plot Deconvolution Results}
\usage{
HOMED_plot(
  beta,
  ref,
  true_prop,
  title = "Deconvolution Results",
  idol = TRUE,
  method = "RPC",
  con = "equality",
  pred_prop = NULL,
  xylim = 0.5
)
}
\arguments{
\item{beta}{A matrix of methylation data (CpG by sample).}

\item{ref}{A reference object containing optimized deconvolution coefficients.}

\item{true_prop}{A matrix or data frame of true cell proportions.}

\item{title}{The title for the plot.}

\item{idol}{Logical; whether to use IDOL-optimized coefficients or not.}

\item{method}{The deconvolution method to use (e.g., `'RPC'`,`'CP'`).}

\item{con}{The constraint for deconvolution (e.g., `'equality'`,`'inequality'`).}

\item{pred_prop}{Optional; a matrix of predicted cell proportions. If not provided, IDOL optimization will be used.}

\item{xylim}{The limits for the x and y axes of the plot.}
}
\value{
A list containing the PCC and CCC data frames, the plot, and predicted proportions.
}
\description{
This function generates a plot of true vs predicted cell proportions for each cell type and computes
the Pearson Correlation Coefficient (PCC) and Concordance Correlation Coefficient (CCC) for each cell type.
The plot shows the correlation between observed and predicted proportions with calculated PCC and CCC values.
}
\examples{
\dontrun{
HOMED_plot(beta_data, reference_data, true_cell_proportions, title = "Deconvolution Results")
}
}
