% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOMED_Helper.R
\name{merge_layers}
\alias{merge_layers}
\title{Merge Cell Type Proportions from Two Layers}
\usage{
merge_layers(layer1_df, layer2_df, cell_types)
}
\arguments{
\item{layer1_df}{A data frame of cell proportions for level 1 cell types.}

\item{layer2_df}{A data frame of cell proportions for level 2 cell types.}

\item{cell_types}{A list where names are cell types and values are vectors of subtypes at the respective levels.}
}
\value{
A data frame with merged and normalized cell type proportions from both layers.
}
\description{
This function merges cell type proportions from two layers (level 1 and level 2) of a hierarchical cell type model.
It normalizes the proportions of the subtypes at level 2 using the proportions at level 1.
}
\examples{
\dontrun{
merged_data <- merge_layers(layer1_data, layer2_data, celltypes)
}
}
